:truck: move redundant code from commands to utils
move ensure_dependencies, ensure_dev_dependencies, guess_toolchain and related structs
