+++ frb_codegen/src/commands.rs
@@ -1,14 +1,12 @@
-use std::collections::HashMap;
 use std::fmt::Write;
 use std::path::Path;
-use std::path::PathBuf;
 use std::process::Command;
 use std::process::Output;
 
 use crate::error::{Error, Result};
+use crate::utils::{ensure_dependencies, ensure_dev_dependencies, guess_toolchain, DartToolchain};
 use cargo_metadata::VersionReq;
 use log::{debug, info, warn};
-use serde::Deserialize;
 
 #[must_use]
 fn call_shell(cmd: &str) -> Output {
@@ -314,7 +312,7 @@ pub fn format_dart(path: &str, line_length: i32) -> Result {
 
 pub fn build_runner(dart_root: &str) -> Result {
     info!("Running build_runner at {}", dart_root);
-    let toolchain = guess_toolchain(dart_root).unwrap();
+    let toolchain = crate::utils::guess_toolchain(dart_root).unwrap();
     let out = if cfg!(windows) {
         call_shell(&format!(
             "cd \"{}\"; {} run build_runner build --delete-conflicting-outputs",
@@ -336,162 +334,4 @@ pub fn build_runner(dart_root: &str) -> Result {
         )));
     }
     Ok(())
-}
-
-#[inline]
-fn read_file(at: &str, filename: &str) -> anyhow::Result<String> {
-    let file = PathBuf::from(at).join(filename);
-    if !file.exists() {
-        return Err(anyhow::Error::msg(format!(
-            "missing {} in {}",
-            filename, at
-        )));
-    }
-    let content = std::fs::read_to_string(file)
-        .map_err(|_| anyhow::Error::msg(format!("unable to read {} in {}", filename, at)))?;
-    Ok(content)
-}
-
-fn guess_toolchain(dart_root: &str) -> anyhow::Result<DartToolchain> {
-    debug!("Guessing toolchain the runner is run into");
-    let lock_file = read_file(dart_root, "pubspec.lock")?;
-    let lock_file: PubspecLock = serde_yaml::from_str(&lock_file).map_err(|_| {
-        anyhow::Error::msg(format!("unable to parse pubspec.lock in {}", dart_root))
-    })?;
-    if lock_file.packages.contains_key("flutter") {
-        return Ok(DartToolchain::Flutter);
-    }
-    Ok(DartToolchain::Dart)
-}
-
-fn ensure_dependencies(dart_root: &str, key: &str) -> anyhow::Result<Option<VersionReq>> {
-    debug!(
-        "Checking presence of {} in dependencies in {}",
-        key, dart_root
-    );
-    let manifest_file = read_file(dart_root, "pubspec.yaml")?;
-    info!("{:#?}", manifest_file);
-    let manifest_file: Pubspec = serde_yaml::from_str(&manifest_file).map_err(|e| {
-        info!("{}", e);
-        anyhow::Error::msg(format!("unable to parse pubspec.yaml in {}", dart_root))
-    })?;
-    let deps = manifest_file.dependencies.unwrap_or_default();
-    let version = deps.get(key);
-    let version = version.map(|v| match v {
-        DependencyVersion::Inline(ref version) => VersionReq::parse(version).unwrap(),
-        DependencyVersion::Multiline { ref version } => VersionReq::parse(
-            version
-                .as_ref()
-                .expect(&format!("please specify a version for {}", key)),
-        )
-        .unwrap(),
-    });
-    info!("found version {:#?} for {}", version, key);
-    Ok(version)
-}
-
-fn ensure_dev_dependencies(dart_root: &str, key: &str) -> anyhow::Result<Option<VersionReq>> {
-    debug!(
-        "Checking presence of {} in dev_dependencies in {}",
-        key, dart_root
-    );
-    let manifest_file = read_file(dart_root, "pubspec.yaml")?;
-    info!("{:#?}", manifest_file);
-    let manifest_file: Pubspec = serde_yaml::from_str(&manifest_file).map_err(|e| {
-        info!("{}", e);
-        anyhow::Error::msg(format!("unable to parse pubspec.yaml in {}", dart_root))
-    })?;
-    let deps = manifest_file.dev_dependencies.unwrap_or_default();
-    let version = deps.get(key);
-    let version = version.map(|v| match v {
-        DependencyVersion::Inline(ref version) => VersionReq::parse(version).unwrap(),
-        DependencyVersion::Multiline { ref version } => VersionReq::parse(
-            version
-                .as_ref()
-                .expect(&format!("please specify a version for {}", key)),
-        )
-        .unwrap(),
-    });
-    info!("found version {:#?} for {}", version, key);
-    Ok(version)
-}
-
-#[derive(Debug, PartialEq)]
-enum DartToolchain {
-    Dart,
-    Flutter,
-}
-
-impl DartToolchain {
-    fn as_run_command(&self) -> &'static str {
-        match self {
-            DartToolchain::Dart => "dart",
-            DartToolchain::Flutter => "flutter pub",
-        }
-    }
-}
-
-#[derive(Debug, Deserialize)]
-struct PubspecLock {
-    pub packages: HashMap<String, serde_yaml::Value>,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(untagged)]
-enum DependencyVersion {
-    Inline(String),
-    Multiline { version: Option<String> },
-}
-
-#[derive(Debug, Deserialize)]
-struct Pubspec {
-    pub dependencies: Option<HashMap<String, DependencyVersion>>,
-    pub dev_dependencies: Option<HashMap<String, DependencyVersion>>,
-}
-
-#[cfg(test)]
-mod tests {
-    use std::path::{Path, PathBuf};
-
-    use crate::commands::{guess_toolchain, DartToolchain};
-    use lazy_static::lazy_static;
-
-    lazy_static! {
-        static ref FRB_EXAMPLES_FOLDER: PathBuf = {
-            PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-                .join("..")
-                .join("frb_example")
-        };
-    }
-
-    fn guess_toolchain_base(path: &Path, expect_toolchain: DartToolchain) {
-        let toolchain = guess_toolchain(&path.to_string_lossy()).expect(&format!(
-            "can get toolchain from {}",
-            path.to_string_lossy()
-        ));
-        assert_eq!(toolchain, expect_toolchain);
-    }
-
-    #[test]
-    fn guess_dart_toolchain() {
-        guess_toolchain_base(
-            FRB_EXAMPLES_FOLDER.join("pure_dart").join("dart").as_path(),
-            DartToolchain::Dart,
-        );
-        guess_toolchain_base(
-            FRB_EXAMPLES_FOLDER
-                .join("pure_dart_multi")
-                .join("dart")
-                .as_path(),
-            DartToolchain::Dart,
-        );
-    }
-
-    #[test]
-    fn guess_flutter_toolchain() {
-        guess_toolchain_base(
-            FRB_EXAMPLES_FOLDER.join("with_flutter").as_path(),
-            DartToolchain::Flutter,
-        );
-    }
 }
